// frontend/src/api/model/registerDTO.ts
/**
 * Generated by orval v7.8.0 🍺
 * Do not edit manually.
 * Rennan API
 * The Rennan API
 * OpenAPI spec version: 1.0.0
 */

export interface RegisterDTO {
  /** Email of the account */
  email: string
  /** Password of the account */
  password: string
  /** Password confirmation */
  password_confirmation: string
  /** Display name of the account */
  name: string
}